feat: giant refactor
