+++ app/routes/__auth/callback.tsx
@@ -10,8 +10,8 @@ import {
 	CircularProgress,
 } from "@chakra-ui/react";
 import { json, LoaderFunction, redirect } from "remix";
-import { authorize, useAuthCallback } from "~app/auth";
-import { url as adminURL } from "~routes/admin";
+import { useAuthCallback, authorize } from "~app/auth";
+import { url as adminURL } from "~routes/__pages/admin/index";
 
 const getLoaderData = async (request: Request) => {
 	if (await authorize(request, { required: false })) throw redirect(adminURL);
